+++ server.js
@@ -12,15 +12,15 @@ require('dotenv').config();
 // console.log(testAccount);
 
 // Creation of the email transporter
-const transporter = nodemailer.createTransport({
-  host: 'smtp.ethereal.email',
-  port: 587,
-  secure: false, // true for 465, false for other ports
-  auth: {
-    user: process.env.EMAIL_USER, // generated ethereal user
-    pass: process.env.EMAIL_PASS // generated ethereal password
-  }
-});
+// const transporter = nodemailer.createTransport({
+//   host: 'smtp.ethereal.email',
+//   port: 587,
+//   secure: false, // true for 465, false for other ports
+//   auth: {
+//     user: process.env.EMAIL_USER, // generated ethereal user
+//     pass: process.env.EMAIL_PASS // generated ethereal password
+//   }
+// });
 
 console.log('> Transporter created');
 
@@ -29,11 +29,19 @@ const PORT = process.env.PORT || (process.env.NODE_ENV === 'test' ? 3001 : 3000)
 app.use(express.urlencoded({ extended: true }));
 
 // middlewares
-app.use((req, res, next) => {
-  req.transporter = transporter;
-  console.log('Middleware transporter');
+// app.use((req, res, next) => {
+//   req.transporter = transporter;
+//   console.log('Middleware transporter');
+//   next();
+// });
+
+app.use( (req,res,next) => {
+  console.log('loggggggggggggggggg');
   next();
-});
+})
+
+
+
 app.use(express.json());
 app.use(cors());
 if (process.env.NODE_ENV !== 'production') {
